Move all configuration-related code to cli/config. Move advertise address check to client for autojoin
